+++ functions/black_list.py
@@ -26,11 +26,11 @@ def add_black_list(obj: BlackListBase) -> bool:
             session.commit()
             return True
 
-def get_user_black_list_id(id: int) -> BlackListBase | bool:
+def get_user_black_list_id(id: int) -> BlackListBase | None | bool:
     with Session() as session:
         try:
             statement = select(BlackListBase).where(BlackListBase.id==int(id))
-            db_object = session.scalars(statement).first()
+            db_object = session.scalars(statement).one_or_none()
 
             _ = db_object.user_black
 
@@ -47,11 +47,11 @@ def get_user_black_list_id(id: int) -> BlackListBase | bool:
 
             return False
 
-def get_user_black_list_user_id(user_id: int) -> BlackListBase | bool:
+def get_user_black_list_user_id(user_id: int) -> BlackListBase | None | bool:
     with Session() as session:
         try:
             statement = select(BlackListBase).where(BlackListBase.id_user==int(user_id))
-            db_object = session.scalars(statement).first()
+            db_object = session.scalars(statement).one_or_none()
 
             _ = db_object.user_black
 
@@ -68,7 +68,7 @@ def get_user_black_list_user_id(user_id: int) -> BlackListBase | bool:
 
             return False
 
-def update_user_black_list(user_id: int, **new_values) -> bool:
+def update_user_black_list(user_id: int, **new_values) -> bool | None:
     """
         Function for updating user in black list.
 
@@ -82,9 +82,12 @@ def update_user_black_list(user_id: int, **new_values) -> bool:
     with Session() as session:
         try:
             user_black_list = get_user_black_list_user_id(user_id)
-            for key, value in new_values.items():
-                if hasattr(user_black_list, key):
-                    setattr(user_black_list, key, value)
+            if user_black_list != None:
+                for key, value in new_values.items():
+                    if hasattr(user_black_list, key):
+                        setattr(user_black_list, key, value)
+            else:
+                return None
 
             session.merge(user_black_list)
 
@@ -105,11 +108,14 @@ def update_user_black_list(user_id: int, **new_values) -> bool:
             session.commit()
             return True
 
-def del_user_black_list(user_id: int) -> bool:
+def del_user_black_list(user_id: int) -> bool | None:
     with Session() as session:
         try:
             user_black_list = get_user_black_list_user_id(user_id)
-            session.delete(user_black_list)
+            if user_black_list != None:
+                session.delete(user_black_list)
+            else:
+                return None
         except Exception as ex:
             session.rollback()
